build: enhance dts gen
