+++ src/vendor-core.ts
