+++ src/vendor.ts
@@ -12,59 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { URL } from 'node:url'
-import {
-  convertPathToPattern,
-  globby,
-  globbySync,
-  globbyStream,
-  generateGlobTasksSync,
-  generateGlobTasks,
-  isGitIgnoredSync,
-  isGitIgnored,
-  isDynamicPattern,
-  type Options as GlobbyOptions,
-} from 'globby'
-import * as yaml from 'yaml'
-import * as _fs from 'fs-extra'
-import _createRequire from 'create-require'
-import { type fetch, AbortController } from 'node-fetch-native'
-
-export * from './core-vendor.js'
-
-export { fetch as nodeFetch } from 'node-fetch-native'
-
-global.AbortController = global.AbortController || AbortController
-
-export const createRequire = _createRequire as unknown as (
-  filename: string | URL
-) => NodeRequire
-
-export const globbyModule = {
-  convertPathToPattern,
-  globby,
-  globbySync,
-  globbyStream,
-  generateGlobTasksSync,
-  generateGlobTasks,
-  isGitIgnoredSync,
-  isGitIgnored,
-  isDynamicPattern,
-}
-
-export const glob = Object.assign(function globby(
-  patterns: string | readonly string[],
-  options?: GlobbyOptions
-) {
-  return globbyModule.globby(patterns, options)
-}, globbyModule) as (typeof globbyModule)['globby'] & typeof globbyModule
-
-export const YAML: {
-  parse(text: string): any
-  stringify(object: any): string
-} = yaml
-
-export const fs: typeof import('fs-extra') = _fs
-
-export { depseekSync as depseek } from 'depseek'
-export { default as minimist } from 'minimist'
+export * from './vendor-core.js'
+export * from './vendor-extra.js'
